requirements.txt
beautifulsoup4==4.12.3
fitz==0.0.1.dev2
Jinja2==3.1.3
openai==1.38.0
pdfminer.six==20231228
Pillow==10.4.0
Requests==2.32.3
